+++ source/Uwmc.Ps.Module.Root.psm1
@@ -8,3 +8,6 @@ Export-ModuleMember -Function Get-OnlinePlayers
 
 # From Uwmc.Ps.Module.PlayerVotes.psm1
 Export-ModuleMember -Function Get-PlayerVotes
+
+# From Uwmc.Ps.Module.Zones.psm1
+Export-ModuleMember -Function Get-Zones